Moved LogItem into items package
