+++ app/src/test/java/ingsw/group1/findmyphone/log/LogItemSpansTest.java
@@ -9,6 +9,7 @@ import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 
 import ingsw.group1.findmyphone.location.GeoPosition;
+import ingsw.group1.findmyphone.log.items.LogItem;
 import ingsw.group1.findmyphone.random.RandomGeoPositionGenerator;
 
 import static org.junit.Assert.assertEquals;
